pyproject.toml
[project]
name = "gammarips-mcp"
version = "1.0.0"
description = "GammaRips MCP Server - Agent-first options trading intelligence"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.6.1",
    "google-api-python-client>=2.0.0",
    "google-cloud-bigquery>=3.11.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-firestore>=2.11.0",
    "httpx>=0.24.0",
    "pandas-market-calendars>=4.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "E402", "E722"]
"src/server.py" = ["E402", "F821", "B904"]  # Closure captures and re-raise patterns
"src/auth/*.py" = ["B904"]
"src/data/*.py" = ["B904"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

